+++ src/pages/Control/hooks/useChangeRow.ts
@@ -1,0 +1,57 @@
+import { useCallback } from 'react'
+import { Atomic } from 'vait'
+import { Base, LayoutInfo } from '../../../core/base'
+import { Matrix } from '../../../core/common'
+import { SearchLayout } from '../../../core/layout'
+import { renderMatrix } from '../../../core/layout/render'
+import { SearchWindow } from '../../../core/layout/window'
+
+type AtomicTasking = ReturnType<typeof Atomic>
+
+export default function useChangeRow(
+  controlProcessing: AtomicTasking,
+  base: Base,
+  layout_info: LayoutInfo,
+  search_layout: SearchLayout | null
+) {
+  return (
+    useCallback(async (type: 'previus' | 'next') => {
+      console.log('changeRow', type, search_layout)
+      if (search_layout === null) {
+        return
+      }
+      return (
+        controlProcessing(async () => {
+          try {
+            search_layout.cancelAllEvent()
+
+            const remainMatrix = [...search_layout.getMatrix()]
+            const latestRow = type === 'next' ? remainMatrix.pop() : remainMatrix.shift()
+
+            let newMatrix: Matrix<SearchWindow>
+
+            if (latestRow === undefined) {
+              throw Error('latestRow is undefined')
+            } else if (type === 'next') {
+              newMatrix = [latestRow, ...remainMatrix]
+            } else {
+              newMatrix = [...remainMatrix, latestRow]
+            }
+
+            await renderMatrix(
+              base,
+              layout_info,
+              newMatrix,
+              type === 'next' ? true : undefined,
+              true
+            )
+
+            search_layout.setMatrix(newMatrix)
+          } finally {
+            search_layout.applyAllEvent()
+          }
+        })
+      )
+    }, [base, controlProcessing, layout_info, search_layout])
+  )
+}
