+++ src/pages/Control/hooks/useChangeRowShortcutKey.ts
@@ -1,0 +1,20 @@
+import { useEffect } from 'react'
+
+export default function useChangeRowShortcutKey(props: {
+  onPressUp: () => void
+  onPressDown: () => void
+}) {
+  useEffect(() => {
+    const handler = (e: KeyboardEvent) => {
+      if (e.key === 'ArrowUp') {
+        e.preventDefault()
+        props.onPressUp()
+      } else if (e.key === 'ArrowDown') {
+        e.preventDefault()
+        props.onPressDown()
+      }
+    }
+    window.addEventListener('keydown', handler, true)
+    return () => window.removeEventListener('keydown', handler, true)
+  }, [props])
+}
